pyproject.toml
[project]
name = "image-styler"
version = "0.1.0"
description = "AI-powered image styling with Streamlit"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "replicate",
    "python-dotenv",
    "pillow",
    "ipykernel>=6.29.5",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "prometheus-client>=0.19.0",
]
